requirements.txt
flask
uvicorn
torch
numpy
pillow
pydantic
facenet-pytorch
scikit-learn
gunicorn
